+++ src/index.ts
@@ -26,6 +26,8 @@ export * from "./jwt.js";
 export * from "./jwks.js";
 export * from "./bootstrap.js";
 export * from "./authkey-session.js";
+export * from "./resolver-session.js";
+export * from "./failover.js";
 export * from "./server-prover.js";
 
 // Admin